Moved CellView and TentacleView out of "model" package
