+++ client/src/main/scala/it/cwmp/client/view/game/CellView.scala
@@ -1,4 +1,4 @@
-package it.cwmp.client.view.game.model
+package it.cwmp.client.view.game
 
 import com.github.tkqubo.colorHash.{ColorHash, Rgb}
 import it.cwmp.client.controller.game.GameConstants
@@ -6,7 +6,6 @@ import it.cwmp.client.model.game.SizingStrategy
 import it.cwmp.client.model.game.impl.{Cell, Point}
 import it.cwmp.client.utils.GeometricUtils
 import it.cwmp.client.view.game.GameViewConstants.{GAME_DEFAULT_FONT_COLOR, GAME_TIME_TEXT_COLOR, RGB_RANGE}
-import it.cwmp.client.view.game.GameViewConstants
 import it.cwmp.client.view.game.drawing.ColoringStrategy
 import javafx.scene.layout._
 import javafx.scene.paint.Color
