+++ client/src/main/scala/it/cwmp/client/view/game/drawing/GameFX.scala
@@ -2,8 +2,7 @@ package it.cwmp.client.view.game.drawing
 
 import akka.actor.ActorRef
 import it.cwmp.client.model.game.impl.{Cell, CellWorld, Point}
-import it.cwmp.client.view.game.GameViewActor
-import it.cwmp.client.view.game.model.{CellView, TentacleView}
+import it.cwmp.client.view.game.{CellView, GameViewActor, TentacleView}
 import it.cwmp.client.view.{FXAlertsController, FXViewController, LayoutRes}
 import javafx.fxml.FXML
 import javafx.scene.canvas.Canvas
